refactor: reorganize tests in directories
